+++ service/src/main/java/org/ehrbase/aql/sql/queryimpl/JsonbEntryQuery.java
@@ -297,8 +297,7 @@ public class JsonbEntryQuery extends ObjectQuery implements IQueryImpl {
 
             dataTypeFromTemplate.evaluate(templateId, referenceItemPathArray);
 
-            if (!jqueryPath.isJsonDataBlock())
-                castTypeAs = dataTypeFromTemplate.getIdentifiedType();
+            castTypeAs = dataTypeFromTemplate.getIdentifiedType();
 
             Field<?> fieldPathItem;
             if (clause.equals(Clause.SELECT)) {
@@ -320,10 +319,8 @@ public class JsonbEntryQuery extends ObjectQuery implements IQueryImpl {
 
             QualifiedAqlField aqlField = new QualifiedAqlField(fieldPathItem,
                                                 dataTypeFromTemplate.getItemType(),
-                                                dataTypeFromTemplate.getItemCategory(),
-                                                jqueryPath.isJsonDataBlock(),
-                                true,
-                                                toAqlPath(itemPathArray));
+                                                dataTypeFromTemplate.getItemCategory()
+                                                );
 
             fieldList.add(aqlField);
         }
@@ -393,7 +390,6 @@ public class JsonbEntryQuery extends ObjectQuery implements IQueryImpl {
 
             Field<?> fieldPathItem = DSL.field(jsqueryPath.toString(), String.class);
             QualifiedAqlField qualifiedAqlField = new QualifiedAqlField(fieldPathItem);
-            qualifiedAqlField.setContainsJqueryPath(true);
             fieldList.add(qualifiedAqlField);
         }
         return new MultiFields(variableDefinition, fieldList, templateId);