+++ service/src/main/java/org/ehrbase/aql/sql/queryimpl/PGType.java
@@ -32,7 +32,7 @@ import static org.jooq.impl.SQLDataType.*;
 public class PGType {
 
     public static final String MAGNITUDE = "magnitude";
-    public static final String VALUE = "value";
+    public static final String COMMA_VALUE = ",value";
     public static final String NUMERATOR = "numerator";
     public static final String DENOMINATOR = "denominator";
     List<String> segmentedPath;
@@ -67,11 +67,11 @@ public class PGType {
                         pgtype = BIGINT;
                     break;
                 case "DV_ORDINAL":
-                    if (StringUtils.endsWith(attribute, VALUE))
+                    if (StringUtils.endsWith(attribute, COMMA_VALUE))
                         pgtype = BIGINT;
                     break;
                 case "DV_BOOLEAN":
-                    if (StringUtils.endsWith(attribute, VALUE))
+                    if (StringUtils.endsWith(attribute, COMMA_VALUE))
                         pgtype = BOOLEAN;
                     break;
                 default: