+++ service/src/main/java/org/ehrbase/aql/sql/queryimpl/value_field/GenericJsonField.java
@@ -23,8 +23,6 @@ public class GenericJsonField extends RMObjectAttribute {
 
     protected Optional<String> jsonPath = Optional.empty();
 
-    private boolean isJsonDataBlock = true; //by default, can be overriden
-
     private static final String ITERATIVE_MARKER = "'"+AQL_NODE_ITERATIVE_MARKER+"'";
 
     public GenericJsonField(FieldResolutionContext fieldContext, JoinSetup joinSetup) {
@@ -85,9 +83,7 @@ public class GenericJsonField extends RMObjectAttribute {
     }
 
     public Field jsonField(String rmType, Object function, TableField... tableFields){
-        fieldContext.setJsonDatablock(isJsonDataBlock);
         fieldContext.setRmType(rmType);
-        //query the json representation of a node and cast the result as TEXT
         Configuration configuration = fieldContext.getContext().configuration();
 
         Field jsonField;
@@ -109,6 +105,7 @@ public class GenericJsonField extends RMObjectAttribute {
         } else
             jsonField = DSL.field(apply(function, tableFields).toString()).cast(String.class);
 
+        //check if the SQL expression contains a set returned in a WHERE clause (implying a lateral join)
         if (jsonField.toString().contains(QueryImplConstants.AQL_NODE_ITERATIVE_FUNCTION) && fieldContext.getClause().equals(IQueryImpl.Clause.WHERE))
             jsonField = DSL.field(DSL.select(jsonField));
 
@@ -165,28 +162,22 @@ public class GenericJsonField extends RMObjectAttribute {
             this.jsonPath = Optional.empty();
             return this;
         }
+
         GenericJsonPath genericJsonPath = new GenericJsonPath(jsonPath);
         this.jsonPath = Optional.of(genericJsonPath.jqueryPath());
         fieldContext.setUsingSetReturningFunction(genericJsonPath.isIterative());
+
         return this;
     }
 
     public GenericJsonField forJsonPath(String root, String jsonPath){
         String actualPath = new AttributePath(root).redux(jsonPath);
-
         return forJsonPath(actualPath);
     }
 
     public GenericJsonField forJsonPath(String[] path){
-        if (GenericJsonPath.isTerminalValue(Arrays.asList(path), path.length - 1))
-            isJsonDataBlock = false;
         this.jsonPath = Optional.of(new JsonbSelect(Arrays.asList(path)).field());
         return this;
     }
 
-
-    public GenericJsonField setJsonDataBlock(boolean jsonDataBlock) {
-        this.isJsonDataBlock = jsonDataBlock;
-        return this;
-    }
 }